requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
pillow
requests
tqdm
kagglehub
flask
flask-cors
